+++ pvlib/spectrum/__init__.py
@@ -1,14 +1,18 @@
 from pvlib.spectrum.spectrl2 import spectrl2  # noqa: F401
-from pvlib.spectrum.mismatch import (  # noqa: F401
+from pvlib.spectrum.spectral_mismatch import (  # noqa: F401
     calc_spectral_mismatch_field,
-    get_am15g,
-    get_reference_spectra,
-    get_example_spectral_response,
     spectral_factor_caballero,
     spectral_factor_firstsolar,
     spectral_factor_sapm,
     spectral_factor_pvspec,
     spectral_factor_jrc,
+)
+from pvlib.spectrum.spectral_irradiance import (  # noqa: F401
+    get_am15g,
+    get_reference_spectra,
+)
+from pvlib.spectrum.spectral_response import (  # noqa: F401
+    get_example_spectral_response,
     sr_to_qe,
     qe_to_sr
 )
