+++ pvlib/spectrum/spectral_mismatch.py
@@ -2,258 +2,13 @@
 The ``mismatch`` module provides functions for spectral mismatch calculations.
 """
 
-import pvlib
-from pvlib._deprecation import deprecated
+from pvlib.spectrum import get_reference_spectra
 from pvlib.tools import normalize_max2one
 import numpy as np
 import pandas as pd
-import scipy.constants
 from scipy.integrate import trapezoid
-from scipy.interpolate import interp1d
 
-from pathlib import Path
 from warnings import warn
-from functools import partial
-
-
-_PLANCK_BY_LIGHT_SPEED_OVER_ELEMENTAL_CHARGE_BY_BILLION = (
-    scipy.constants.speed_of_light
-    * scipy.constants.Planck
-    / scipy.constants.elementary_charge
-    * 1e9
-)
-
-
-def get_example_spectral_response(wavelength=None):
-    '''
-    Generate a generic smooth spectral response (SR) for tests and experiments.
-
-    Parameters
-    ----------
-    wavelength: 1-D sequence of numeric, optional
-        Wavelengths at which spectral response values are generated.
-        By default ``wavelength`` is from 280 to 1200 in 5 nm intervals. [nm]
-
-    Returns
-    -------
-    spectral_response : pandas.Series
-        The relative spectral response indexed by ``wavelength`` in nm. [-]
-
-    Notes
-    -----
-    This spectral response is based on measurements taken on a c-Si cell.
-    A small number of points near the measured curve are used to define
-    a cubic spline having no undue oscillations, as shown in [1]_.  The spline
-    can be interpolated at arbitrary wavelengths to produce a continuous,
-    smooth curve , which makes it suitable for experimenting with spectral
-    data of different resolutions.
-
-    References
-    ----------
-    .. [1] Driesse, Anton, and Stein, Joshua. "Global Normal Spectral
-       Irradiance in Albuquerque: a One-Year Open Dataset for PV Research".
-       United States 2020. :doi:`10.2172/1814068`.
-    '''
-    # Contributed by Anton Driesse (@adriesse), PV Performance Labs. Aug. 2022
-
-    SR_DATA = np.array([[290, 0.00],
-                        [350, 0.27],
-                        [400, 0.37],
-                        [500, 0.52],
-                        [650, 0.71],
-                        [800, 0.88],
-                        [900, 0.97],
-                        [950, 1.00],
-                        [1000, 0.93],
-                        [1050, 0.58],
-                        [1100, 0.21],
-                        [1150, 0.05],
-                        [1190, 0.00]]).transpose()
-
-    if wavelength is None:
-        resolution = 5.0
-        wavelength = np.arange(280, 1200 + resolution, resolution)
-
-    interpolator = interp1d(SR_DATA[0], SR_DATA[1],
-                            kind='cubic',
-                            bounds_error=False,
-                            fill_value=0.0,
-                            copy=False,
-                            assume_sorted=True)
-
-    sr = pd.Series(data=interpolator(wavelength), index=wavelength)
-
-    sr.index.name = 'wavelength'
-    sr.name = 'spectral_response'
-
-    return sr
-
-
-@deprecated(
-    since="0.11",
-    removal="0.12",
-    name="pvlib.spectrum.get_am15g",
-    alternative="pvlib.spectrum.get_reference_spectra",
-    addendum=(
-        "The new function reads more data. Use it with "
-        + "standard='ASTM G173-03' and extract the 'global' column."
-    ),
-)
-def get_am15g(wavelength=None):
-    r"""
-    Read the ASTM G173-03 AM1.5 global spectrum on a 37-degree tilted surface,
-    optionally interpolated to the specified wavelength(s).
-
-    Global (tilted) irradiance includes direct and diffuse irradiance from sky
-    and ground reflections, and is more formally called hemispherical
-    irradiance (on a tilted surface).  In the context of photovoltaic systems
-    the irradiance on a flat receiver is frequently called plane-of-array (POA)
-    irradiance.
-
-    Parameters
-    ----------
-    wavelength: 1-D sequence of numeric, optional
-        Wavelengths at which the spectrum is interpolated.
-        By default the 2002 wavelengths of the standard are returned. [nm].
-
-    Returns
-    -------
-    am15g: pandas.Series
-        The AM1.5g standard spectrum indexed by ``wavelength``. [W/(m²nm)].
-
-    Notes
-    -----
-    If ``wavelength`` is specified this function uses linear interpolation.
-
-    If the values in ``wavelength`` are too widely spaced, the integral of the
-    spectrum may deviate from the standard value of 1000.37 W/m².
-
-    The values in the data file provided with pvlib-python are copied from an
-    Excel file distributed by NREL, which is found here:
-    https://www.nrel.gov/grid/solar-resource/assets/data/astmg173.xls
-
-    More information about reference spectra is found here:
-    https://www.nrel.gov/grid/solar-resource/spectra-am1.5.html
-
-    See Also
-    --------
-    pvlib.spectrum.get_reference_spectra : reads also the direct and
-      extraterrestrial components of the spectrum.
-
-    References
-    ----------
-    .. [1] ASTM "G173-03 Standard Tables for Reference Solar Spectral
-       Irradiances: Direct Normal and Hemispherical on 37° Tilted Surface."
-    """  # noqa: E501
-    # Contributed by Anton Driesse (@adriesse), PV Performance Labs. Aug. 2022
-    # modified by @echedey-ls, as a wrapper of spectrum.get_reference_spectra
-    standard = get_reference_spectra(wavelength, standard="ASTM G173-03")
-    return standard["global"]
-
-
-def get_reference_spectra(wavelengths=None, standard="ASTM G173-03"):
-    r"""
-    Read a standard spectrum specified by ``standard``, optionally
-    interpolated to the specified wavelength(s).
-
-    Defaults to ``ASTM G173-03`` AM1.5 standard [1]_, which returns
-    ``extraterrestrial``, ``global`` and ``direct`` spectrum on a 37-degree
-    tilted surface, optionally interpolated to the specified wavelength(s).
-
-    Parameters
-    ----------
-    wavelengths : numeric, optional
-        Wavelengths at which the spectrum is interpolated. [nm].
-        If not provided, the original wavelengths from the specified standard
-        are used. Values outside that range are filled with zeros.
-
-    standard : str, default "ASTM G173-03"
-        The reference standard to be read. Only the reference
-        ``"ASTM G173-03"`` is available at the moment.
-
-    Returns
-    -------
-    standard_spectra : pandas.DataFrame
-        The standard spectrum by ``wavelength [nm]``. [W/(m²nm)].
-        Column names are ``extraterrestrial``, ``direct`` and ``global``.
-
-    Notes
-    -----
-    If ``wavelength`` is specified, linear interpolation is used.
-
-    If the values in ``wavelength`` are too widely spaced, the integral of each
-    spectrum may deviate from its standard value.
-    For global spectra, it is about 1000.37 W/m².
-
-    The values of the ASTM G173-03 provided with pvlib-python are copied from
-    an Excel file distributed by NREL, which is found here [2]_:
-    https://www.nrel.gov/grid/solar-resource/assets/data/astmg173.xls
-
-    Examples
-    --------
-    >>> from pvlib import spectrum
-    >>> am15 = spectrum.get_reference_spectra()
-    >>> am15_extraterrestrial, am15_global, am15_direct = \
-    >>>     am15['extraterrestrial'], am15['global'], am15['direct']
-    >>> print(am15.head())
-                extraterrestrial        global        direct
-    wavelength
-    280.0                  0.082  4.730900e-23  2.536100e-26
-    280.5                  0.099  1.230700e-21  1.091700e-24
-    281.0                  0.150  5.689500e-21  6.125300e-24
-    281.5                  0.212  1.566200e-19  2.747900e-22
-    282.0                  0.267  1.194600e-18  2.834600e-21
-
-    >>> am15 = spectrum.get_reference_spectra([300, 500, 800, 1100])
-    >>> print(am15)
-                extraterrestrial   global    direct
-    wavelength
-    300                  0.45794  0.00102  0.000456
-    500                  1.91600  1.54510  1.339100
-    800                  1.12480  1.07250  0.988590
-    1100                 0.60000  0.48577  0.461130
-
-    References
-    ----------
-    .. [1] ASTM "G173-03 Standard Tables for Reference Solar Spectral
-       Irradiances: Direct Normal and Hemispherical on 37° Tilted Surface."
-    .. [2] “Reference Air Mass 1.5 Spectra,” www.nrel.gov.
-       https://www.nrel.gov/grid/solar-resource/spectra-am1.5.html
-    """  # Contributed by Echedey Luis, inspired by Anton Driesse (get_am15g)
-    SPECTRA_FILES = {
-        "ASTM G173-03": "ASTMG173.csv",
-    }
-    pvlib_datapath = Path(pvlib.__path__[0]) / "data"
-
-    try:
-        filepath = pvlib_datapath / SPECTRA_FILES[standard]
-    except KeyError:
-        raise ValueError(
-            f"Invalid standard identifier '{standard}'. Available "
-            + "identifiers are: "
-            + ", ".join(SPECTRA_FILES.keys())
-        )
-
-    standard = pd.read_csv(
-        filepath,
-        header=1,  # expect first line of description, then column names
-        index_col=0,  # first column is "wavelength"
-        dtype=float,
-    )
-
-    if wavelengths is not None:
-        interpolator = partial(
-            np.interp, xp=standard.index, left=0.0, right=0.0
-        )
-        standard = pd.DataFrame(
-            index=wavelengths,
-            data={
-                col: interpolator(x=wavelengths, fp=standard[col])
-                for col in standard.columns
-            },
-        )
-
-    return standard
 
 
 def calc_spectral_mismatch_field(sr, e_sun, e_ref=None):
@@ -953,108 +708,6 @@ def spectral_factor_jrc(airmass, clearsky_index, module_type=None,
         + coeff[2] * (airmass - 1.5)
     )
     return mismatch
-
-
-def sr_to_qe(sr, wavelength=None, normalize=False):
-    """
-    Convert spectral responsivities to quantum efficiencies.
-    If ``wavelength`` is not provided, the spectral responsivity ``sr`` must be
-    a :py:class:`pandas.Series` or :py:class:`pandas.DataFrame`, with the
-    wavelengths in the index.
-
-    Provide wavelengths in nanometers, [nm].
-
-    Conversion is described in [1]_.
-
-    .. versionadded:: 0.11.0
-
-    Parameters
-    ----------
-    sr : numeric, pandas.Series or pandas.DataFrame
-        Spectral response, [A/W].
-        Index must be the wavelength in nanometers, [nm].
-
-    wavelength : numeric, optional
-        Points where spectral response is measured, in nanometers, [nm].
-
-    normalize : bool, default False
-        If True, the quantum efficiency is normalized so that the maximum value
-        is 1.
-        For ``pandas.DataFrame``, normalization is done for each column.
-        For 2D arrays, normalization is done for each sub-array.
-
-    Returns
-    -------
-    quantum_efficiency : numeric, same type as ``sr``
-        Quantum efficiency, in the interval [0, 1].
-
-    Notes
-    -----
-    - If ``sr`` is of type ``pandas.Series`` or ``pandas.DataFrame``,
-      column names will remain unchanged in the returned object.
-    - If ``wavelength`` is provided it will be used independently of the
-      datatype of ``sr``.
-
-    Examples
-    --------
-    >>> import numpy as np
-    >>> import pandas as pd
-    >>> from pvlib import spectrum
-    >>> wavelengths = np.array([350, 550, 750])
-    >>> spectral_response = np.array([0.25, 0.40, 0.57])
-    >>> quantum_efficiency = spectrum.sr_to_qe(spectral_response, wavelengths)
-    >>> print(quantum_efficiency)
-    array([0.88560142, 0.90170326, 0.94227991])
-
-    >>> spectral_response_series = pd.Series(spectral_response, index=wavelengths, name="dataset")
-    >>> qe = spectrum.sr_to_qe(spectral_response_series)
-    >>> print(qe)
-    350    0.885601
-    550    0.901703
-    750    0.942280
-    Name: dataset, dtype: float64
-
-    >>> qe = spectrum.sr_to_qe(spectral_response_series, normalize=True)
-    >>> print(qe)
-    350    0.939850
-    550    0.956938
-    750    1.000000
-    Name: dataset, dtype: float64
-
-    References
-    ----------
-    .. [1] “Spectral Response,” PV Performance Modeling Collaborative (PVPMC).
-        https://pvpmc.sandia.gov/modeling-guide/2-dc-module-iv/effective-irradiance/spectral-response/
-    .. [2] “Spectral Response | PVEducation,” www.pveducation.org.
-        https://www.pveducation.org/pvcdrom/solar-cell-operation/spectral-response
-
-    See Also
-    --------
-    pvlib.spectrum.qe_to_sr
-    """  # noqa: E501
-    if wavelength is None:
-        if hasattr(sr, "index"):  # true for pandas objects
-            # use reference to index values instead of index alone so
-            # sr / wavelength returns a series with the same name
-            wavelength = sr.index.array
-        else:
-            raise TypeError(
-                "'sr' must have an '.index' attribute"
-                + " or 'wavelength' must be provided"
-            )
-    quantum_efficiency = (
-        sr
-        / wavelength
-        * _PLANCK_BY_LIGHT_SPEED_OVER_ELEMENTAL_CHARGE_BY_BILLION
-    )
-
-    if normalize:
-        quantum_efficiency = normalize_max2one(quantum_efficiency)
-
-    return quantum_efficiency
-
-
-def qe_to_sr(qe, wavelength=None, normalize=False):
     """
     Convert quantum efficiencies to spectral responsivities.
     If ``wavelength`` is not provided, the quantum efficiency ``qe`` must be
